refactor(Cards): Cards 를 state 안으로 구조 이동
