+++ src/main/java/blackjack/domain/status/Blackjack.java
@@ -1,7 +1,5 @@
 package blackjack.domain.status;
 
-import blackjack.domain.participant.Cards;
-
 public class Blackjack extends Finished {
     public Blackjack(Cards cards) {
         super(cards);
