+++ src/main/java/blackjack/domain/status/Bust.java
@@ -1,7 +1,5 @@
 package blackjack.domain.status;
 
-import blackjack.domain.participant.Cards;
-
 public class Bust extends Finished {
     public Bust(Cards cards) {
         super(cards);
