+++ src/main/java/blackjack/domain/status/StartRound.java
@@ -1,7 +1,6 @@
 package blackjack.domain.status;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.participant.Cards;
 
 public class StartRound {
     public static State draw(final Card first, final Card second) {
