+++ src/main/java/blackjack/domain/status/Stay.java
@@ -1,7 +1,5 @@
 package blackjack.domain.status;
 
-import blackjack.domain.participant.Cards;
-
 public class Stay extends Finished{
     public Stay(Cards cards) {
         super(cards);
